# Pre-ncc compiled packages to prevent re-nccing unnecessarily