--- a/.prettierignore
+++ b/.prettierignore
@@ -2,3 +2,4 @@ node_modules
 **/.next/**
 **/_next/**
 **/dist/**
+packages/next/compiled/**