--- a/.prettierignore_staged
+++ b/.prettierignore_staged
@@ -1,3 +1,4 @@
 **/.next/**
 **/_next/**
 **/dist/**
+packages/next/compiled/**/*