--- a/lint-staged.config.js
+++ b/lint-staged.config.js
@@ -7,7 +7,7 @@ module.exports = {
       .map(filename => `"${isWin ? filename : escape([filename])}"`)
       .join(' ')
     return [
-      `prettier --with-node-modules --ignore-path='./.prettierignore_staged' --write ${escapedFileNames}`,
+      `prettier --with-node-modules --ignore-path .prettierignore_staged --write ${escapedFileNames}`,
       `eslint --no-ignore --max-warnings=0 --fix ${filenames
         .map(f => `"${f}"`)
         .join(' ')}`,
@@ -19,7 +19,7 @@ module.exports = {
       .map(filename => `"${isWin ? filename : escape([filename])}"`)
       .join(' ')
     return [
-      `prettier --with-node-modules --ignore-path='./.prettierignore_staged' --write ${escapedFileNames}`,
+      `prettier --with-node-modules --ignore-path .prettierignore_staged --write ${escapedFileNames}`,
       `git add ${escapedFileNames}`,
     ]
   },