--- a/packages/next/compiled/amphtml-validator/package.json
+++ b/packages/next/compiled/amphtml-validator/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "amphtml-validator",
+  "main": "index.js",
+  "author": "The AMP HTML Authors",
+  "license": "Apache-2.0"
+}