--- a/packages/next/compiled/arg/package.json
+++ b/packages/next/compiled/arg/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "arg",
+  "main": "index.js",
+  "author": "Josh Junon <junon@zeit.co>",
+  "license": "MIT"
+}