--- a/packages/next/compiled/async-retry/package.json
+++ b/packages/next/compiled/async-retry/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "async-retry",
+  "main": "index.js",
+  "license": "MIT"
+}