This attempts to speed up local development and pre-publish build times by pre-nccing and committing the compiled packages to our repo since they should only change when we update/add new packages to be ncc'd and don't need to be re-ncc'd every time we are building `next` locally for development or every time before publish 

**Before this optimization**
Dev build startup: 23 - 25 seconds
Prod build: 52 - 55 seconds

**After this optimization**
Dev build startup: 5 - 7 seconds
Prod build: 35 - 37 seconds

<details>
<summary>Before</summary>

<img width="927" alt="before-prod" src="https://user-images.githubusercontent.com/22380829/78187403-f0c7c700-7433-11ea-8e14-a9ce1ee7f2bd.png">
<img width="927" alt="before-dev" src="https://user-images.githubusercontent.com/22380829/78187405-f2918a80-7433-11ea-8923-42c638a0bef0.png">
</details>

<details>
<summary>After</summary>

<img width="927" alt="after-dev" src="https://user-images.githubusercontent.com/22380829/78187441-02a96a00-7434-11ea-9d03-302f312a3dbd.png">
<img width="927" alt="after-prod" src="https://user-images.githubusercontent.com/22380829/78187448-05a45a80-7434-11ea-840c-b627ed0206b5.png">
</details>